lava_callback: Move simple URL validator to base Service
This patch extracts function from "lava_callback" to keep it in a more
generic package. In future it might be moved to e.g. "helpers" one but
for now base Service will do.
